+++ internal/adapter/in_memory_db/position_db/delete.go
@@ -4,10 +4,11 @@ import (
 	"fmt"
 )
 
-func (pr *PositionRepo) DeletePosition(id int) error {
+func (pr *PositionRepo) DeletePosition(id int64) error {
 	if _, exists := pr.positionsMap[id]; !exists {
 		return fmt.Errorf("position does not exist")
 	}
 	delete(pr.positionsMap, id)
+
 	return nil
 }
